Refactor twenty days ago to registration paradigm.
Move DNS test fixtures into test_caa.
