+++ test/helpers.py
@@ -11,6 +11,9 @@ import shutil
 import atexit
 import subprocess
 
+import challtestsrv
+
+challSrv = challtestsrv.ChallTestServer()
 tempdir = tempfile.mkdtemp()
 
 @atexit.register
@@ -161,3 +164,18 @@ def verify_revocation(cert_file, issuer_file, url):
     if not re.search(": revoked", verify_output):
         print verify_output
         raise Exception("OCSP response wasn't 'revoked'")
+
+twenty_days_ago_functions = [ ]
+
+def register_twenty_days_ago(f):
+    """Register a function to be run during "setup_twenty_days_ago." This allows
+       test cases to define their own custom setup.
+    """
+    twenty_days_ago_functions.append(f)
+
+def setup_twenty_days_ago():
+    """Do any setup that needs to happen 20 day in the past, for tests that
+       will run in the 'present'.
+    """
+    for f in twenty_days_ago_functions:
+        f()
