+++ test/v1_integration.py
@@ -425,6 +425,16 @@ def test_revoke_by_account():
     verify_akamai_purge()
     return 0
 
+caa_authzs = []
+@register_twenty_days_ago
+def caa_setup():
+    # Issue a certificate with the clock set back, and save the authzs to check
+    # later that they are valid (200). They should however require rechecking for
+    # CAA purposes.
+    _, authzs = auth_and_issue(["recheck.good-caa-reserved.com"], client=caa_client)
+    for a in authzs:
+        caa_authzs.append(a)
+
 def test_caa():
     """Request issuance for two CAA domains, one where we are permitted and one where we are not.
        Two further sub-domains have restricted validationmethods.
@@ -437,6 +447,22 @@ def test_caa():
             raise Exception("Unexpected response for CAA authz: ",
                 response.status_code)
 
+    goodCAA = "happy-hacker-ca.invalid"
+    badCAA = "sad-hacker-ca.invalid"
+
+    caa_account_uri = caa_client.account.uri if caa_client is not None else None
+    caa_records = [
+        {"domain": "bad-caa-reserved.com", "value": badCAA},
+        {"domain": "good-caa-reserved.com", "value": goodCAA},
+        {"domain": "accounturi.good-caa-reserved.com", "value":"{0}; accounturi={1}".format(goodCAA, caa_account_uri)},
+        {"domain": "recheck.good-caa-reserved.com", "value":badCAA},
+        {"domain": "dns-01-only.good-caa-reserved.com", "value": "{0}; validationmethods=dns-01".format(goodCAA)},
+        {"domain": "http-01-only.good-caa-reserved.com", "value": "{0}; validationmethods=http-01".format(goodCAA)},
+        {"domain": "dns-01-or-http01.good-caa-reserved.com", "value": "{0}; validationmethods=dns-01,http-01".format(goodCAA)},
+    ]
+    for policy in caa_records:
+        challSrv.add_caa_issue(policy["domain"], policy["value"])
+
     # We include a random domain so we don't hit the "exact match" rate limit
     # when testing locally with a persistent database.
     auth_and_issue(["good-caa-reserved.com"])
@@ -445,6 +471,7 @@ def test_caa():
     # now be denied due to CAA.
     chisel.expect_problem("urn:acme:error:caa", lambda: chisel.issue(caa_client, caa_authzs))
 
+    challSrv.add_caa_issue({"domain": "bad-caa-reserved.com", "value": badCAA})
     chisel.expect_problem("urn:acme:error:caa",
         lambda: auth_and_issue(["bad-caa-reserved.com"]))
